+++ core/src/utils/mod.rs
@@ -1,5 +1,3 @@
-pub mod sync;
-
 use alloc::vec::Vec;
 use core::{
     fmt::Debug,
@@ -8,6 +6,7 @@ use core::{
 
 // RE-EXPORTS
 // ================================================================================================
+pub use miden_core_diagnostics::sync;
 pub use miden_crypto::utils::{
     ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable, SliceReader,
     collections, uninit_vector,